+++ packages/core/src/indexer/index.ts
@@ -1,1 +1,2 @@
 export * from "./indexer.js";
+export * from "./quality.js";